requirements.txt
aspy.yaml==0.3.0
cached-property==1.3.0
identify==1.0.5
ipython
nodeenv==1.2.0
pre-commit==0.16.3
py==1.4.34
pytest==3.2.1
PyYAML==3.12
requests
six==1.10.0
virtualenv==15.1.0
